README.md
This project is a Membership Management System initially built with HTML, PHP, and MySQL for fitness centers, featuring user authentication and data storage. Later, it was upgraded to a production-ready Single Page Application using ReactJS for the frontend and Spring Boot for the backend, significantly improving scalability, user engagement, and administrative efficiency.